producer-rssfeed/requirement.txt
kafka-python
pandas
numpy
requests
pymongo
gnewsclient